# src/BSDCPolls.BFF/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BSDCPolls BFF",
    "version": "v1"
  },
  "paths": {
...
    "/api/polls": {
      "get": {
        "tags": ["Polls"],
        "operationId": "Polls_GetFeed",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/PollStatus" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollFeedResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "Bearer": [] }]
      },
      "post": {
        "tags": ["Polls"],
        "operationId": "Polls_Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePollRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollDetailResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/polls/{pollUid}": {
      "get": {
        "tags": ["Polls"],
        "operationId": "Polls_GetByUid",
        "parameters": [
          { "name": "pollUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollDetailResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/polls/{pollUid}/status": {
      "patch": {
        "tags": ["Polls"],
        "operationId": "Polls_ChangeStatus",
        "parameters": [
          { "name": "pollUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePollStatusRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollDetailResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/polls/{pollUid}/questions": {
      "post": {
        "tags": ["Polls"],
        "operationId": "Polls_AddQuestion",
        "parameters": [
          { "name": "pollUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddPollQuestionRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollQuestionResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/polls/{pollUid}/questions/{questionUid}/push": {
      "post": {
        "tags": ["Polls"],
        "operationId": "Polls_PushQuestion",
        "parameters": [
          { "name": "pollUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "questionUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollQuestionResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
...
    "/api/polls/{pollUid}/results": {
      "get": {
        "tags": ["Polls"],
        "operationId": "Polls_GetResults",
        "parameters": [
          { "name": "pollUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollResultsResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PollStatus": {
        "type": "string",
        "enum": ["Draft", "Active", "Closed"]
      },
      "CreatePollRequest": {
        "type": "object",
        "required": ["title", "isPublic"],
        "properties": {
          "title": { "type": "string", "maxLength": 200 },
          "isPublic": { "type": "boolean" }
        }
      },
      "ChangePollStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "$ref": "#/components/schemas/PollStatus" }
        }
      },
      "AnswerOptionInput": {
        "type": "object",
        "required": ["text", "orderIndex"],
        "properties": {
          "text": { "type": "string", "maxLength": 200 },
          "orderIndex": { "type": "integer" }
        }
      },
      "AddPollQuestionRequest": {
        "type": "object",
        "required": ["text", "options", "pushImmediately"],
        "properties": {
          "text": { "type": "string", "maxLength": 500 },
          "options": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AnswerOptionInput" },
            "minItems": 2,
            "maxItems": 10
          },
          "pushImmediately": { "type": "boolean" }
        }
      },
      "PollAnswerOptionResponse": {
        "type": "object",
        "required": ["optionUid", "text", "orderIndex"],
        "properties": {
          "optionUid": { "type": "string", "format": "uuid" },
          "text": { "type": "string" },
          "orderIndex": { "type": "integer" }
        }
      },
      "PollQuestionResponse": {
        "type": "object",
        "required": ["questionUid", "text", "orderIndex", "options"],
        "properties": {
          "questionUid": { "type": "string", "format": "uuid" },
          "text": { "type": "string" },
          "orderIndex": { "type": "integer" },
          "pushedAt": { "type": "string", "format": "date-time", "nullable": true },
          "options": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollAnswerOptionResponse" }
          }
        }
      },
      "PollDetailResponse": {
        "type": "object",
        "required": ["pollUid", "title", "isPublic", "status", "createdOn", "isCreator", "questions"],
        "properties": {
          "pollUid": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "isPublic": { "type": "boolean" },
          "status": { "$ref": "#/components/schemas/PollStatus" },
          "createdOn": { "type": "string", "format": "date-time" },
          "isCreator": { "type": "boolean" },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollQuestionResponse" }
          }
        }
      },
      "PollFeedItem": {
        "type": "object",
        "required": ["pollUid", "title", "isPublic", "status", "creatorUsername", "questionCount", "createdOn"],
        "properties": {
          "pollUid": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "isPublic": { "type": "boolean" },
          "status": { "$ref": "#/components/schemas/PollStatus" },
          "creatorUsername": { "type": "string" },
          "questionCount": { "type": "integer" },
          "createdOn": { "type": "string", "format": "date-time" },
          "invitedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "PollFeedResponse": {
        "type": "object",
        "required": ["items", "totalCount", "page", "pageSize"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollFeedItem" }
          },
          "totalCount": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" }
        }
      },
      "PollResultsOptionResponse": {
        "type": "object",
        "required": ["optionUid", "text", "voteCount", "percentage"],
        "properties": {
          "optionUid": { "type": "string", "format": "uuid" },
          "text": { "type": "string" },
          "voteCount": { "type": "integer" },
          "percentage": { "type": "number" }
        }
      },
      "PollResultsQuestionResponse": {
        "type": "object",
        "required": ["questionUid", "text", "totalVotes", "options"],
        "properties": {
          "questionUid": { "type": "string", "format": "uuid" },
          "text": { "type": "string" },
          "pushedAt": { "type": "string", "format": "date-time", "nullable": true },
          "totalVotes": { "type": "integer" },
          "options": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollResultsOptionResponse" }
          }
        }
      },
...
      "PollResultsResponse": {
        "type": "object",
        "required": ["pollUid", "title", "status", "questions"],
        "properties": {
          "pollUid": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "status": { "$ref": "#/components/schemas/PollStatus" },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollResultsQuestionResponse" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}